refactor(): split related functions into separate files
Split related functions into separate files as dom-driver.js
Was very large and had multiple concerns.
