+++ src/index.js
@@ -1,8 +1,8 @@
 import h from 'snabbdom/h'
 import thunk from 'snabbdom/thunk'
 const hh = require(`hyperscript-helpers`)(h)
-import makeDOMDriver from './dom-driver'
+import makeDomDriver from './makeDomDriver'
 
 import assign from 'fast.js/object/assign'
 
-module.exports = assign({makeDOMDriver, h, thunk}, hh)
+module.exports = assign({makeDomDriver, h, thunk}, hh)
